refactor: extract lit datamodule factory module
